refactor(reports): finish phase 3 - codes/ holds calculation only
The wall's report tables (_compile_wall_shear_dicts, _compile_results_wall_shear,
160 lines) were the last presentation left inside codes/. They moved to
reports/walls.py, next to the wall's views and Word reports, and
_check_shear_ACI_318_19_wall returns None - the wall decides when to build a
report, exactly as the beam does.

  codes/ACI_318_19_wall.py  512 -> 343

The boundary test now walks EVERY module under codes/ instead of the two beam
ones, and fails on any _initialize_dicts_* or _compile_* function. Verified by
adding one back to the wall module and watching it fail, then restoring.

Two mypy errors surfaced because reports/walls.py is type-checked while the code
module it came from is not: the wall's concrete is typed as the base Concrete,
so phi_v and lambda_factor were unknown. Narrowed through a small _aci() helper
that says why the cast is safe - the check raises TypeError for any other
concrete before a report can exist.

Phase 3 is complete. codes/ is equations and orchestration, elements are
geometry and thin delegation, presentation is one place:

  mento/reports/                1929 lines
  mento/plots/                   991 lines
  mento/codes/  (calculation)   2276 lines

Suite 965 to 969 passed, coverage still 100%, mypy clean across 45 modules.
